income is float

+++ hippo/hippo/AccountBook.h
@@ -17,6 +17,6 @@
 @property (nonatomic) NSString* use;
 @property (nonatomic) NSDate* date;
 
-- (NSInteger)income;
+- (float)income;
 
 @end
